Refactored tests as well
